Create README.md
This is the first project that I was assigned to do where I was supposed to use text cells to mention the most commonly used programming languages in point form and then create a table containing the headlines: Intermediate devices, End devices, and Media with examples in each column. Then, I used code cells to print my name and by utilizing the input() command, I took the name of a user to greet them "Hello" and used int(input()) to accept two numbers to print their sum.